pyproject.toml
[tool.isort]
known_third_party = "discord"

[tool.poetry]
name = "telescope2"
version = "0.10.4a1"
description = ""
authors = ["tonyzbf <tonyzbf@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.6"
Django = "^3.2.5"
aiodns = "^3.0.0"
aiofiles = "^0.7.0"
aiohttp = "^3.7.4"
attrs = "^21.2.0"
beautifulsoup4 = "^4.9.3"
cchardet = "^2.1.7"
"discord.py" = "^1.7.3"
geopy = "^2.2.0"
inflect = "^5.3.0"
Jinja2 = "^3.0.1"
lxml = "^4.6.3"
Markdown = "^3.3.4"
nltk = "^3.6.2"
pendulum = "^2.1.2"
Pillow = "^8.3.1"
psutil = "^5.8.0"
PyJWT = "^2.1.0"
pylibmc = "^1.6.1"
python-decouple = "^3.4"
simplejson = "^3.17.3"
termcolor = "^1.1.0"
timezonefinder = "^5.2.0"
aiojobs = "^0.3.0"
daphne = "^3.0.2"
django-cacheops = "^6.0"
django-extensions = "^3.1.3"
django-polymorphic = "^3.0.0"
django-redis = "^5.0.0"
django-timezone-field = "^4.2.1"
fuzzywuzzy = "^0.18.0"
graphene-django = "3.0.0b7"
more-itertools = "^8.8.0"
python-Levenshtein = "^0.12.2"
redis = "^3.5.3"
Unidecode = "^1.2.0"
Faker = "^8.10.2"
django-csp = "^3.7"
django-manifest-loader = "^1.0.0"
toml = "^0.10.2"
dateparser = "^1.0.0"
asyncstdlib = "^3.9.2"
Pint = "^0.17"
APScheduler = "^3.7.0"
duckdb = "^0.2.9"
pandas = "^1.3.3"
"discord.jinja" = {git = "https://git.dougdistrict.xyz/discord.jinja.git", rev = "master"}
"duckcord.py" = {git = "https://git.dougdistrict.xyz/duckcord.py.git", rev = "master"}
arrow = "^1.2.1"

[tool.poetry.dev-dependencies]
logging_tree = "^1.9"
pipdeptree = "^2.0.0"
snakeviz = "^2.1.0"
dpytest = "^0.5.3"

[build-system]
requires = ["poetry>=1.1.7"]
build-backend = "poetry.masonry.api"
